+++ modules/gltf/src/index.js
@@ -11,7 +11,6 @@ export {default as GLBParser} from './glb/glb-parser';
 export {default as GLBBuilder} from './glb/glb-builder';
 
 export {
-  ATTRIBUTE_TYPE_TO_COMPONENTS,
-  ATTRIBUTE_COMPONENT_TYPE_TO_BYTE_SIZE,
-  ATTRIBUTE_COMPONENT_TYPE_TO_ARRAY
-} from './utils/gltf-type-utils';
+  KHR_DRACO_MESH_COMPRESSION,
+  UBER_POINT_CLOUD_EXTENSION
+} from './gltf/gltf-constants';