+++ modules/gltf/test/glb/glb-parser.spec.js
@@ -10,7 +10,7 @@ test('GLBParser#parse', t => {
   const arrayBuffer = builder.encodeAsGLB();
 
   const parser = new GLBParser();
-  parser.parse(arrayBuffer);
+  parser.parseSync(arrayBuffer);
   const json = parser.getApplicationData('test');
   t.ok(json.nested.typedArray instanceof Float32Array, 'Float32Array present');