+++ docs/whats-new.md
@@ -8,8 +8,9 @@ This module is currently in unofficial "soft pre-release" stage. You are welcome
 ## v0.8
 
 * `fetchFile` - New function supports fetch API under browser and Node.js
-* `glTFLoader` - improvements, supports async loading of linked resources and base64 encoded buffer.
-* `glTFParser` - API changes: `parse()` is now async and supports linked resource loading, and a separate `parseSync()` method is provided for applications that depend on the old sync behavior and do not need the new features.
+* `GLTFLoader` - improvements, supports async loading of linked resources and base64 encoded buffer.
+* `GLTFParser` - API changes: `parse()` is now async and supports linked resource loading, and a separate `parseSync()` method is provided for applications that depend on the old sync behavior and do not need the new features.
+* `GLBBuilder`,`GLTFBuilder` - `nopack` option renamed to `packTypedArrays`. It must now explicitly be set to `true` to enable packing.
 
 
 ## v0.7